refactor gitea impl
