+++ interface/forges/gitea/gitea.py
@@ -25,126 +25,20 @@ from requests.auth import HTTPBasicAuth
 from rfc3339 import rfc3339
 from dynaconf import settings
 
-from .base import (
+from interface.forges.base import (
     Forge,
     F_D_REPOSITORY_NOT_FOUND,
     F_D_FORGE_FORBIDDEN_OPERATION,
     F_D_REPOSITORY_EXISTS,
     F_D_INVALID_ISSUE_URL,
 )
-from .payload import CreateIssue, RepositoryInfo, CreatePullrequest
-from .notifications import Notification, NotificationResp, Comment
-from .notifications import ISSUE, PULL, COMMIT, REPOSITORY
+from interface.forges.payload import CreateIssue, RepositoryInfo, CreatePullrequest
+from interface.forges.notifications import Notification, NotificationResp, Comment
+from interface.forges.notifications import ISSUE, PULL, COMMIT, REPOSITORY
 from interface.error import F_D_FORGE_UNKNOWN_ERROR, Error
 from interface.utils import trim_url, clean_url, get_rand
 
-
-class ParseCSRFGiteaForm(HTMLParser):
-    token: str = None
-
-    def handle_starttag(self, tag: str, attrs: (str, str)):
-        if self.token:
-            return
-
-        if tag != "input":
-            return
-
-        token = None
-        for (index, (k, v)) in enumerate(attrs):
-            if k == "value":
-                token = v
-
-            if all([k == "name", v == "_csrf"]):
-                if token:
-                    self.token = token
-                    return
-                for (inner_index, (nk, nv)) in enumerate(attrs, start=index):
-                    if nk == "value":
-                        self.token = nv
-                        return
-
-
-class HTMLClient:
-    session: Session
-
-    def __init__(self):
-        self.host = urlparse(clean_url(settings.GITEA.host))
-        if all([self.host.scheme != "http", self.host.scheme != "https"]):
-            print(self.host.scheme)
-            raise Exception("scheme should be either http or https")
-        self.session = Session()
-        self.login()
-        print(f"constructor {self.session.cookies}")
-
-    @staticmethod
-    def get_csrf_token(page: str) -> str:
-        parser = ParseCSRFGiteaForm()
-        parser.feed(page)
-        csrf = parser.token
-        return csrf
-
-    def get_url(self, path: str) -> str:
-        return urlunparse((self.host.scheme, self.host.netloc, path, "", "", ""))
-
-    def login(self):
-        url = self.get_url("/user/login")
-        resp = self.session.get(url)
-        if resp.status_code != 200:
-            print(resp.status_code, resp.text)
-            raise Exception(resp.status_code)
-
-        csrf = self.get_csrf_token(resp.text)
-        payload = {
-            "_csrf": csrf,
-            "user_name": settings.GITEA.username,
-            "password": settings.GITEA.password,
-            "remember": "on",
-        }
-        resp = self.session.post(url, data=payload, allow_redirects=False)
-        print(f"login {self.session.cookies}")
-        if resp.status_code == 302:
-            return
-
-        raise Exception(
-            f"[ERROR] Authentication failed. status code {resp.status_code}"
-        )
-
-    def fork(self, repo_id: int, repo_name: str, uid: int):
-        url = self.get_url(f"/repo/fork/{repo_id}")
-        payload = {
-            "uid": uid,
-            "repo_name": repo_name,
-            "description": "",
-        }
-
-        def __inner(payload, count: int = 0):
-            print(payload)
-            resp = self.session.get(url)
-            if resp.status_code != 200:
-                # Have to see source code for possible errors(wrong password? user doesn't exist?)
-                print(resp.status_code, resp.text)
-                raise Exception(resp.status_code)
-
-            csrf = self.get_csrf_token(resp.text)
-            payload["_csrf"] = csrf
-            print(f"payload in gitea: {payload}")
-            resp = self.session.post(url, data=payload, allow_redirects=False)
-            if resp.status_code == 302:
-                return self.get_url(resp.headers["location"])
-            if (
-                '<a href="/user/sign_up">Need an account? Register now.</a>'
-                in resp.text
-            ):
-                print("logging in")
-                self.login()
-                count += 1
-                return __inner(count)
-            print(resp.text)
-            raise Exception(
-                f"[ERROR]: CSRF forking repo_id: {repo_id} status: {resp.status_code}"
-            )
-
-        return __inner(payload)
+from .html_client import HTMLClient
 
 
 class Gitea(Forge):
@@ -262,6 +156,12 @@ class Gitea(Forge):
         subject = n["subject"]
         notification_type = subject["type"]
 
+        # 2021-12-25: REPOSITORY type notification is only sent out when a
+        # repository transfer request is made and the recipient has to confirm
+        # it --- irrelevant for us
+        if notification_type == REPOSITORY:
+            return None
+
         last_read = n["updated_at"]
         rn = Notification(
             updated_at=last_read,
@@ -272,10 +172,7 @@ class Gitea(Forge):
             repo_url=n["repository"]["html_url"],
         )
 
-        if notification_type == REPOSITORY:
-            print(n)
-
-        elif notification_type == PULL:
+        if notification_type == PULL:
             rn.pr_url = requests.request("GET", subject["url"]).json()["html_url"]
 
             rn.upstream = n["repository"]["description"]
@@ -327,8 +224,9 @@ class Gitea(Forge):
         for n in notifications:
             # rn: Repository Notification
             rn = self._into_notification(n)
-            last_read = rn.updated_at
-            resp.append(rn)
+            if rn:
+                last_read = rn.updated_at
+                resp.append(rn)
         return NotificationResp(notifications=resp, last_read=date_parse(last_read))
 
     def create_pull_request(self, owner: str, repo: str, pr: CreatePullrequest):
